+++ package.json
@@ -22,7 +22,7 @@
     "assemble:android:prod": "cd android && ./gradlew assembleRelease && cd ..",
     "assemble:android:stage": "cd android && ./gradlew assembleDebug && cd ..",
     "prebuild": "npm run env && node ./tools/git-version.js && npm run build:packages",
-    "prepare": "husky && ts-patch install && npm run env && npm run shared-schema:build && npm run plugin-api:build",
+    "prepare": "husky && ts-patch install && npm run env && npm run shared-schema:build && npm run sync-core:build && npm run plugin-api:build",
     "build": "npm run buildAllElectron:noTests:prod",
     "build:packages": "node ./packages/build-packages.js",
     "buildAllElectron:noTests:prod": "npm run lint && npm run buildFrontend:prod:es6 && npm run electron:build",
@@ -155,6 +155,7 @@
     "clean:translations": "node ./tools/clean-translations.js",
     "plugin-api:build": "cd packages/plugin-api && npm run build",
     "shared-schema:build": "cd packages/shared-schema && npm run build",
+    "sync-core:build": "cd packages/sync-core && npm run build",
     "plugin-api:build:watch": "cd packages/plugin-api && npm run build:watch",
     "vite-plugin:build": "cd packages/vite-plugin && npm run build",
     "plugins:build": "npm run plugin-api:build && npm run vite-plugin:build && cd packages/plugin-dev && npm run build:all",
